+++ com.unity.netcode.gameobjects/Tests/Runtime/NetworkVariableTests.cs
@@ -9,7 +9,7 @@ using Random = UnityEngine.Random;
 
 namespace Unity.Netcode.RuntimeTests
 {
-    public struct TestStruct : INetworkSerializable
+    public struct TestStruct : INetworkSerializable, IEquatable<TestStruct>
     {
         public uint SomeInt;
         public bool SomeBool;
@@ -28,6 +28,24 @@ namespace Unity.Netcode.RuntimeTests
             }
             serializer.SerializeValue(ref SomeInt);
             serializer.SerializeValue(ref SomeBool);
+        }
+
+        public bool Equals(TestStruct other)
+        {
+            return SomeInt == other.SomeInt && SomeBool == other.SomeBool;
+        }
+
+        public override bool Equals(object obj)
+        {
+            return obj is TestStruct other && Equals(other);
+        }
+
+        public override int GetHashCode()
+        {
+            unchecked
+            {
+                return ((int)SomeInt * 397) ^ SomeBool.GetHashCode();
+            }
         }
     }
 
@@ -50,6 +68,7 @@ namespace Unity.Netcode.RuntimeTests
         }
 
         public readonly NetworkVariable<TestStruct> TheStruct = new NetworkVariable<TestStruct>();
+        public readonly NetworkList<TestStruct> TheListOfStructs = new NetworkList<TestStruct>();
 
         public bool ListDelegateTriggered;
 
@@ -172,12 +191,11 @@ namespace Unity.Netcode.RuntimeTests
         [UnityTest]
         public IEnumerator AllNetworkVariableTypes([Values(true, false)] bool useHost)
         {
-            NetworkManager server;
             // Create, instantiate, and host
             // This would normally go in Setup, but since every other test but this one
             //  uses MultiInstanceHelper, and it does its own NetworkManager setup / teardown,
             //  for now we put this within this one test until we migrate it to MIH
-            Assert.IsTrue(NetworkManagerHelper.StartNetworkManager(out server, useHost ? NetworkManagerHelper.NetworkManagerOperatingMode.Host : NetworkManagerHelper.NetworkManagerOperatingMode.Server));
+            Assert.IsTrue(NetworkManagerHelper.StartNetworkManager(out NetworkManager server, useHost ? NetworkManagerHelper.NetworkManagerOperatingMode.Host : NetworkManagerHelper.NetworkManagerOperatingMode.Server));
 
             Assert.IsTrue(server.IsHost == useHost, $"{nameof(useHost)} does not match the server.IsHost value!");
 
@@ -354,6 +372,22 @@ namespace Unity.Netcode.RuntimeTests
         }
 
         [UnityTest]
+        public IEnumerator TestListOfINetworkSerializableCallsNetworkSerialize([Values(true, false)] bool useHost)
+        {
+            yield return InitializeServerAndClients(useHost);
+            yield return MultiInstanceHelpers.RunAndWaitForCondition(
+                () =>
+                {
+                    TestStruct.NetworkSerializeCalledOnWrite = false;
+                    TestStruct.NetworkSerializeCalledOnRead = false;
+                    m_Player1OnServer.TheListOfStructs.Add(new TestStruct() { SomeInt = k_TestUInt, SomeBool = false });
+                    m_Player1OnServer.TheListOfStructs.SetDirty(true);
+                },
+                () => TestStruct.NetworkSerializeCalledOnWrite && TestStruct.NetworkSerializeCalledOnRead
+            );
+        }
+
+        [UnityTest]
         public IEnumerator TestNetworkVariableStruct([Values(true, false)] bool useHost)
         {
             yield return InitializeServerAndClients(useHost);
@@ -379,11 +413,7 @@ namespace Unity.Netcode.RuntimeTests
             TestStruct.NetworkSerializeCalledOnRead = false;
             m_Player1OnServer.TheStruct.Value = new TestStruct() { SomeInt = k_TestUInt, SomeBool = false };
 
-            bool VerifyCallback()
-            {
-                return TestStruct.NetworkSerializeCalledOnWrite &&
-                    TestStruct.NetworkSerializeCalledOnRead;
-            }
+            static bool VerifyCallback() => TestStruct.NetworkSerializeCalledOnWrite && TestStruct.NetworkSerializeCalledOnRead;
 
             // Wait for the client-side to notify it is finished initializing and spawning.
             yield return WaitForConditionOrTimeOut(VerifyCallback);
@@ -579,12 +609,14 @@ namespace Unity.Netcode.RuntimeTests
             m_NetworkListTestState = networkListTestState;
             m_Player1OnServer = player1OnServer;
             m_Player1OnClient1 = player1OnClient1;
-            m_StateFunctions = new Dictionary<NetworkListTestStates, Func<bool>>();
-            m_StateFunctions.Add(NetworkListTestStates.Add, OnAdd);
-            m_StateFunctions.Add(NetworkListTestStates.ContainsLarge, OnContainsLarge);
-            m_StateFunctions.Add(NetworkListTestStates.Contains, OnContains);
-            m_StateFunctions.Add(NetworkListTestStates.VerifyData, OnVerifyData);
-            m_StateFunctions.Add(NetworkListTestStates.IndexOf, OnIndexOf);
+            m_StateFunctions = new Dictionary<NetworkListTestStates, Func<bool>>
+            {
+                { NetworkListTestStates.Add, OnAdd },
+                { NetworkListTestStates.ContainsLarge, OnContainsLarge },
+                { NetworkListTestStates.Contains, OnContains },
+                { NetworkListTestStates.VerifyData, OnVerifyData },
+                { NetworkListTestStates.IndexOf, OnIndexOf }
+            };
 
             if (networkListTestState == NetworkListTestStates.ContainsLarge)
             {